add setup instructions, requirements files for dependencies, config file that can be modified without changing versioned code
