--- src/main/cv/README.md
@@ -1,1 +1,0 @@
-# CV Code
