+++ src/main/cv/python/.gitignore
@@ -1,1 +1,2 @@
 venv/
+.env
