+++ src/main/cv/python/StreamDriverCam.py
@@ -1,21 +1,13 @@
+from config import *
 import cv2
 import time
 from http.server import BaseHTTPRequestHandler, HTTPServer
 import socket
-from linuxpy.video.device import Device
 
-PORT = 1181
-LISTEN_IP =  '0.0.0.0' # 0.0.0.0 means accessible from any ip
-CAM_USB_ID = "usb-xhci-hcd.0-2" # based on port of raspberry pi (top right)
-
-def find_camera_index(usbId):
-    index = 0
-    while index < 4: # there can't be more than 4 cameras, I think
-        with Device.from_id(index) as cam:
-            if cam.info.bus_info == usbId:
-                return index
-
-index = find_camera_index(CAM_USB_ID)
+if ON_RPI:
+    index = find_camera_index(CAM_USB_ID)
+else:
+    index = DRIVER_CAM_INDEX
 camera = cv2.VideoCapture(index)
 class MJPEGStreamHandler(BaseHTTPRequestHandler):
     def do_GET(self):
@@ -47,9 +39,9 @@ class MJPEGStreamHandler(BaseHTTPRequestHandler):
 
 
 def StreamDriverCam():
-    server_address = (LISTEN_IP, PORT) 
+    server_address = (DRIVER_CAM_LISTEN_IP, DRIVER_CAM_LISTEN_PORT) 
     httpd = HTTPServer(server_address, MJPEGStreamHandler)
-    print(f'Streaming video at http://{LISTEN_IP}:{PORT}/stream.mjpg')
+    print(f'Streaming video at http://{DRIVER_CAM_LISTEN_IP}:{DRIVER_CAM_LISTEN_PORT}/stream.mjpg')
     httpd.serve_forever()
 
 StreamDriverCam()
