+++ src/main/cv/python/requirements_rpi.txt
@@ -1,0 +1,5 @@
+opencv-python
+pupil_apriltags
+linuxpy
+--extra-index-url https://wpilib.jfrog.io/artifactory/api/pypi/wpilib-python-release-2024/simple
+pyntcore
